@@ -1,51 +1,57 @@
 {
-  "name": "bootstrap-seed",
-  "version": "1.0.0",
+  "name": "bootstap-seed",
+  "version": "1.0.1",
   "license": "MIT",
   "scripts": {
+    "ng": "ng",
     "start": "ng serve",
-    "lint": "tslint \"src/**/*.ts\"",
+    "build": "ng build --prod",
+    "cdoc": "compodoc -p src\\tsconfig.app.json",
     "test": "ng test",
-    "pree2e": "webdriver-manager update",
-    "e2e": "protractor",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
     "build-electron": "ng build --base-href . && copy src\\electron\\* dist",
     "electron": "npm run build-electron && electron dist\\electron.js",
     "jquery": "copy node_modules\\jquery\\dist\\jquery.min.js src\\assets\\js"
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~4.1",
-    "@angular/compiler": "~4.1",
-    "@angular/core": "~4.1",
-    "@angular/forms": "~4.1",
-    "@angular/http": "~4.1",
-    "@angular/platform-browser": "~4.1",
-    "@angular/platform-browser-dynamic": "~4.1",
-    "@angular/router": "~4.1",
-    "bootstrap": "^3.3.0",
-    "core-js": "^2.4.1",
-    "jquery": "^3.0.0",
-    "rxjs": "^5.1.0",
-    "zone.js": "^0.8.4"
+    "@angular/animations": "~8.2.4",
+    "@angular/common": "~8.2.4",
+    "@angular/compiler": "~8.2.4",
+    "@angular/core": "~8.2.4",
+    "@angular/forms": "~8.2.4",
+    "@angular/platform-browser": "~8.2.4",
+    "@angular/platform-browser-dynamic": "~8.2.4",
+    "@angular/router": "~8.2.4",
+    "bootstrap": "^4.3.1",
+    "core-js": "^2.5.7",
+    "jquery": "^3.3.1",
+    "popper.js": "^1.14.7",
+    "rxjs": "~6.4.0",
+    "tslib": "^1.10.0",
+    "zone.js": "~0.9.1"
   },
   "devDependencies": {
-    "@angular/cli": "^1.7.4",
-    "@angular/compiler-cli": "~4.1",
-    "@types/jasmine": "2.5.38",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~2.0.0",
-    "electron": "1.8.4",
-    "jasmine-core": "~2.5.2",
-    "jasmine-spec-reporter": "~3.2.0",
-    "karma": "~1.4.1",
-    "karma-chrome-launcher": "~2.0.0",
-    "karma-cli": "~1.0.1",
-    "karma-jasmine": "~1.1.0",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "karma-coverage-istanbul-reporter": "^0.2.0",
-    "protractor": "~5.1.0",
-    "ts-node": "~2.0.0",
-    "tslint": "~4.5.0",
-    "typescript": "~2.2.0"
+    "@angular-devkit/build-angular": "~0.803.2",
+    "@angular/cli": "~8.3.2",
+    "@angular/compiler-cli": "~8.2.4",
+    "@angular/language-service": "~8.2.4",
+    "@types/node": "~8.9.4",
+    "@types/jasmine": "~3.3.8",
+    "@types/jasminewd2": "~2.0.3",
+    "codelyzer": "^5.0.0",
+    "electron": "1.8.8",
+    "jasmine-core": "~3.4.0",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~4.1.0",
+    "karma-chrome-launcher": "~2.2.0",
+    "karma-coverage-istanbul-reporter": "~2.0.1",
+    "karma-jasmine": "~2.0.1",
+    "karma-jasmine-html-reporter": "^1.4.0",
+    "protractor": "~5.4.0",
+    "ts-node": "~7.0.0",
+    "tslint": "~5.15.0",
+    "typescript": "~3.5.3"
   }
 }
