@@ -1,7 +1,7 @@
 import { BrowserModule } from '@angular/platform-browser';
 import { NgModule }      from '@angular/core';
 import { FormsModule }   from '@angular/forms';
-import { HttpModule }    from '@angular/http';
+import {HttpClientModule} from '@angular/common/http';
 
 import { AboutModule }     from './about/about.module';
 import { AppComponent }    from './app.component';
@@ -19,7 +19,7 @@ import { SharedModule }    from './shared/shared.module';
     BrowserModule,
     FormsModule,
     HomeModule,
-    HttpModule,
+    HttpClientModule,
     SharedModule.forRoot()
   ],
   providers: [],
